+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.2.0 - Alpha
+* Initial index option added for `VideosSourceController.fromUrlList`` constructor
+* VideoControllerConfiguration of [media_kit](https://pub.dev/packages/media_kit) exported to final user
+
 ## 0.1.0 - Alpha
-* TODO: Initial release.
+* Initial release.